+++ system/lib/ex_esdb/emitter_system.ex
@@ -11,11 +11,14 @@ defmodule ExESDB.EmitterSystem do
   use Supervisor
 
   alias ExESDB.Themes, as: Themes
+  alias ExESDB.StoreNaming
 
   @impl true
-  def init(_opts) do
+  def init(opts) do
+    store_id = StoreNaming.extract_store_id(opts)
+    
     children = [
-      {PartitionSupervisor, child_spec: DynamicSupervisor, name: ExESDB.EmitterPools}
+      {PartitionSupervisor, child_spec: DynamicSupervisor, name: StoreNaming.partition_name(ExESDB.EmitterPools, store_id)}
     ]
 
     # Use :one_for_one - simple structure
@@ -31,12 +34,17 @@ defmodule ExESDB.EmitterSystem do
   end
 
   def start_link(opts) do
-    Supervisor.start_link(__MODULE__, opts, name: __MODULE__)
+    store_id = StoreNaming.extract_store_id(opts)
+    name = StoreNaming.genserver_name(__MODULE__, store_id)
+    
+    Supervisor.start_link(__MODULE__, opts, name: name)
   end
 
   def child_spec(opts) do
+    store_id = StoreNaming.extract_store_id(opts)
+    
     %{
-      id: __MODULE__,
+      id: StoreNaming.child_spec_id(__MODULE__, store_id),
       start: {__MODULE__, :start_link, [opts]},
       # Only restart when abnormal
       restart: :transient,