+++ system/lib/ex_esdb/store_system.ex
@@ -16,6 +16,7 @@ defmodule ExESDB.StoreSystem do
   use Supervisor
 
   alias ExESDB.Themes, as: Themes
+  alias ExESDB.StoreNaming
 
   @impl true
   def init(opts) do
@@ -40,12 +41,17 @@ defmodule ExESDB.StoreSystem do
   end
 
   def start_link(opts) do
-    Supervisor.start_link(__MODULE__, opts, name: __MODULE__)
+    store_id = StoreNaming.extract_store_id(opts)
+    name = StoreNaming.genserver_name(__MODULE__, store_id)
+    
+    Supervisor.start_link(__MODULE__, opts, name: name)
   end
 
   def child_spec(opts) do
+    store_id = StoreNaming.extract_store_id(opts)
+    
     %{
-      id: __MODULE__,
+      id: StoreNaming.child_spec_id(__MODULE__, store_id),
       start: {__MODULE__, :start_link, [opts]},
       restart: :permanent,
       shutdown: :infinity,