+++ system/lib/ex_esdb/subscriptions_tracker.ex
@@ -15,6 +15,7 @@ defmodule ExESDB.SubscriptionsTracker do
   alias ExESDB.Emitters, as: Emitters
   alias ExESDB.StoreCluster, as: StoreCluster
   alias ExESDB.Themes, as: Themes
+  alias ExESDB.StoreNaming
 
   ########### HANDLE_INFO ###########
   @impl GenServer
@@ -122,20 +123,26 @@ defmodule ExESDB.SubscriptionsTracker do
     :ok
   end
 
-  def child_spec(opts),
-    do: %{
-      id: __MODULE__,
+  def child_spec(opts) do
+    store_id = StoreNaming.extract_store_id(opts)
+    
+    %{
+      id: StoreNaming.child_spec_id(__MODULE__, store_id),
       start: {__MODULE__, :start_link, [opts]},
       type: :worker,
       restart: :permanent,
       shutdown: 5000
     }
+  end
 
-  def start_link(opts),
-    do:
-      GenServer.start_link(
-        __MODULE__,
-        opts,
-        name: __MODULE__
-      )
+  def start_link(opts) do
+    store_id = StoreNaming.extract_store_id(opts)
+    name = StoreNaming.genserver_name(__MODULE__, store_id)
+    
+    GenServer.start_link(
+      __MODULE__,
+      opts,
+      name: name
+    )
+  end
 end